requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
